ForwardAnalysis: Do not declare classes in `__init__.py`
As a side effect: do not expose content unrelated to this module from
it...
